requirements.txt
dahuffman
python-rle
deflate
arithmetic_compressor
scipy


# usage: pip install -r requirements.txt